+++ packages/ui-causal-graph-editor/src/graph-viewer/causal-graph-editor.tsx
@@ -20,8 +20,9 @@ import { useEffect, useMemo, useState } from 'react';
 import * as React from 'react';
 import { GetReferenceClientRect } from 'tippy.js';
 
-import { useTheme } from '@darajs/styled-components';
+import styled, { useTheme } from '@darajs/styled-components';
 import { Tooltip } from '@darajs/ui-components';
+import { Notification, NotificationPayload } from '@darajs/ui-notifications';
 import { Status, useUpdateEffect } from '@darajs/ui-utils';
 import { ConfirmationModal } from '@darajs/ui-widgets';
 
@@ -64,6 +65,32 @@ import useDragMode from '../shared/use-drag-mode';
 import { useEdgeConstraintEncoder } from '../shared/use-edge-encoder';
 import useIterateEdges from './utils/use-iterate-edges';
 import useIterateNodes from './utils/use-iterate-nodes';
+
+const NotificationWrapper = styled.div`
+    position: relative;
+    z-index: 100;
+    top: calc(100% - 8rem);
+    left: calc(100% - 23rem);
+
+    height: 0;
+
+    div {
+        align-items: center;
+        height: 7rem;
+        padding: 0.75rem;
+    }
+
+    span {
+        flex-grow: 1;
+
+        -webkit-line-clamp: 3;
+    }
+
+    h2 {
+        align-self: baseline;
+        height: 1.6rem;
+    }
+`;
 
 export interface CausalGraphEditorProps extends Settings {
     /** Optional additional legends to show */
@@ -124,6 +151,11 @@ function CausalGraphEditor(props: CausalGraphEditorProps): JSX.Element {
         () => props.editorMode ?? (isDag(state.graph) ? EditorMode.DEFAULT : EditorMode.PAG_VIEWER)
     );
 
+    const [error, setError] = useState(null);
+    const handleError = (e: NotificationPayload): void => {
+        setError(e);
+    };
+
     const {
         getCenterPosition,
         useEngineEvent,
@@ -141,6 +173,7 @@ function CausalGraphEditor(props: CausalGraphEditorProps): JSX.Element {
         props.editable,
         editorMode,
         props.initialConstraints,
+        handleError,
         props.processEdgeStyle,
         props.zoomThresholds
     );
@@ -643,6 +676,11 @@ function CausalGraphEditor(props: CausalGraphEditorProps): JSX.Element {
                                 )}
                             </GraphContext.Provider>
                         </Overlay>
+                        {error && (
+                            <NotificationWrapper>
+                                <Notification notification={error} onDismiss={() => setError(null)} />
+                            </NotificationWrapper>
+                        )}
                         <div ref={canvasParentRef} style={{ height: '100%', width: '100%' }} />
                         <Tooltip
                             content={tooltipContent}