+++ packages/ui-causal-graph-editor/src/graph-viewer/graph-editor.stories.tsx
@@ -168,6 +168,19 @@ PlanarHorizontal.args = {
     graphLayout: PlanarLayout.Builder.build(),
 };
 
+export const PlanarTiers = Template.bind({});
+const planarLayout = PlanarLayout.Builder.build();
+planarLayout.tiers = { group: 'meta.test', order_nodes_by: 'order' };
+// planarLayout.tiers = nodeTiersList;
+// planarLayout.tiers = { group: 'meta.group', rank: ['a', 'b', 'c', 'd', 'e'] };
+
+PlanarTiers.args = {
+    editable: true,
+    graphData: nodeTiersCausalGraph,
+    // graphData: FRAUD,
+    graphLayout: planarLayout,
+};
+
 export const Spring = Template.bind({});
 Spring.args = {
     editable: true,