+++ packages/ui-notifications/src/index.tsx
@@ -17,3 +17,4 @@
 export { NotificationPayload } from './notification-payload';
 export { default as NotificationWrapper, NotificationContext } from './notification-wrapper';
 export { useNotifications, UseNotificationsInterface } from './use-notifications';
+export { default as Notification } from './notification';