+++ packages/ui-notifications/src/notification.tsx
@@ -83,8 +83,8 @@ const Body = styled.span<NotificationBodyProps>`
     display: -webkit-box; /* stylelint-disable-line value-no-vendor-prefix */
 
     font-weight: 400;
-    text-overflow: ellipsis;
     line-height: 1.25rem;
+    text-overflow: ellipsis;
 
     -webkit-box-orient: vertical;
     -webkit-line-clamp: ${(props) => (props.moreDetailsShown ? 1 : 2)};
@@ -101,12 +101,16 @@ const Icon = styled.div<NotificationIconsProps>`
 `;
 
 const MoreDetailsButton = styled.button`
-    color: ${() => theme.colors.grey4};
-    border: none;
-    background: transparent;
     cursor: pointer;
+
     display: flex;
+
     padding-left: 0;
+
+    color: ${() => theme.colors.grey4};
+
+    background: transparent;
+    border: none;
     :hover {
         text-decoration: underline;
     }
@@ -131,7 +135,7 @@ export interface NotificationProps {
     /** The notification to render */
     notification: NotificationPayload;
     /** Handler to trigger the dismissal of the notification */
-    onDismiss: (key: string) => void | Promise<void>;
+    onDismiss?: (key: string) => void | Promise<void>;
     /** Optional handler to trigger when the more details button is clicked */
     onMoreDetailsClick?: (
         e: React.MouseEvent<HTMLButtonElement, MouseEvent>,